cell-grapher/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cell-grapher"
version = "0.1.0"
description = "A tool for creating region adjacency graphs and analyzing T1 transitions in microscopy images"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Tianyi Cao", email = "ctyjackcao@outlook.com" }
]
dependencies = [
    "numpy",
    "matplotlib",
    "scikit-image",
    "networkx",
    "scipy",
    "pyyaml",
]

[project.scripts]
cell-grapher = "cell_grapher.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cell_grapher"]
